% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.r
\name{define}
\alias{define}
\title{Load a bunch of dependencies by filename}
\usage{
define(..., envir = parent.env(topenv()), local)
}
\arguments{
\item{...}{see examples.}

\item{envir}{environment. The parent environment to use when calling
\code{base::source} to fetch dependencies.}

\item{local}{logical. If \code{TRUE} and \code{envir} is missing,
it will set \code{envir = parent.frame()}.}
}
\description{
This is useful for reducing pollution in the global namespace,
and not loading multiple files twice unnecessarily.
}
\examples{
\dontrun{
helper_fn <- define('some/dir/helper_fn')
define(c('some/dir/helper_fn', 'some/other_dir/library_fn'), function(helper_fn, library_fn) { ... }
helper_fns <<- define('some/dir/helper_fn1', 'some/otherdir/helper_fn2')
helper_fns[[1]]('do something'); helper_fns[[2]]('do something else')
}
}

