% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/load_dependency.r
\name{load_dependency}
\alias{load_dependency}
\title{Load a bunch of dependencies by filename}
\usage{
load_dependency(dep)
}
\arguments{
\item{dep}{Name of dependency, e.g., relative filename (without .r)}
}
\description{
Load a bunch of dependencies by filename
}
\examples{
\dontrun{
  helper <- load_dependency('path/to/helper')
}
}

