% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_package.r
\name{load_package}
\alias{load_package}
\title{Load a package and install it if necessary}
\usage{
load_package(name, verbose = FALSE)
}
\arguments{
\item{name}{character. Name of package.}

\item{verbose}{logical. Whether or not to announce each installation.}
}
\description{
Load a package and install it if necessary
}
\examples{
\dontrun{
}
}

