% Generated by roxygen2 (4.0.1): do not edit by hand
\name{packages}
\alias{packages}
\title{Load a bunch of packages}
\usage{
packages(...)
}
\arguments{
\item{...}{List of packages}
}
\description{
Load a bunch of packages
}
\examples{
\dontrun{
packages('glmnet', 'caret)
packages('glmnet caret')
packages(list('glmnet', 'caret'), c('e1071', 'parallel multicore'), 'stringr')
}
}

