% Generated by roxygen2 (4.0.1): do not edit by hand
\name{define}
\alias{define}
\title{Load a bunch of dependencies by filename}
\usage{
define(...)
}
\arguments{
\item{...}{see examples.}
}
\description{
This is useful for reducing pollution in the global namespace,
and not loading multiple files twice unnecessarily.
}
\examples{
\dontrun{
helper_fn <- define('some/dir/helper_fn')
define(c('some/dir/helper_fn', 'some/other_dir/library_fn'), function(helper_fn, library_fn) { ... }
helper_fns <<- define('some/dir/helper_fn1', 'some/otherdir/helper_fn2')
helper_fns[[1]]('do something'); helper_fns[[2]]('do something else')
}
}

