% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_dependency}
\alias{load_dependency}
\title{Load a bunch of dependencies by filename}
\arguments{
\item{dep}{Name of dependency, e.g., relative filename (without .r)}
}
\description{
Load a bunch of dependencies by filename
}
\examples{
\dontrun{
  helper <- load_dependency('path/to/helper')
}
}

