% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/load_package.r
\name{load_package}
\alias{load_package}
\title{Load a package and install it if necessary}
\usage{
load_package(name)
}
\arguments{
\item{name}{Name of package}
}
\description{
Load a package and install it if necessary
}
\examples{
\dontrun{
load_package('glmnet')
}
}

