% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package.r
\name{load_package}
\alias{load_package}
\title{Load a package and install it if necessary}
\usage{
load_package(name, verbose = FALSE)
}
\arguments{
\item{name}{character. Name of package.}

\item{verbose}{logical. Whether or not to announce each installation.}
}
\description{
Load a package and install it if necessary
}
\examples{
\dontrun{
load_package("glmnet")
load_package("robertzk/Ramd")
load_package("robertzk/Ramd@v0.3")   # Can load from versions
load_package("robertzk/Ramd@0.3")
load_package("robertzk/Ramd@fbe1aa0e36df289b27881d077635352e6debdbc1")  # Can load from refs
load_package(list("FeiYeYe/xgboost", subdir = "R-package"))  # Can load from subdirectories
}
}
