% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packages.r
\name{packages}
\alias{packages}
\title{Load a bunch of packages}
\usage{
packages(..., verbose = FALSE)
}
\arguments{
\item{...}{List of packages}

\item{verbose}{logical. Whether or not to announce installations and
package startup messages.}
}
\description{
Load a bunch of packages
}
\examples{
\dontrun{
packages("glmnet", "caret")
packages("glmnet caret")  # Can just separate with a space
packages("robertzk/Ramd")  # Can install from GitHub
packages("robertzk/Ramd", "hadley/dplyr", "peterhurford/batchman")
packages("robertzk/Ramd", "glmnet")  # Can install from both CRAN and GitHub
}
}

